+++ build.gradle
@@ -46,7 +46,7 @@ def jsonProviders = [
          group: 'org.jetbrains.kotlinx', name: 'kotlinx-serialization-json', version: '1.7.3'],
         [id   : 'jakartaJson', additionalVersions: [],
          group: 'jakarta.json', name: 'jakarta.json-api', version: '2.1.3'],
-        [id   : 'jettison', additionalVersions: [],
+        [id   : 'jettison', additionalVersions: ['1.5.0'],
          group: 'org.codehaus.jettison', name: 'jettison', version: '1.5.4'],
         [id   : 'jsonSmart', additionalVersions: ['1.1'],
          group: 'net.minidev', name: 'json-smart', version: '2.5.1'],
